+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- hadolint: keep [hadolint](https://github.com/hadolint/hadolint) `sync`ed and `update`d
+
 - vim: keep [neovim](https://github.com/neovim/neovim) `sync`ed and `update`d if detected
 
 ## [0.16.0] - 2018-07-13